disable icd10 import

+++ db/seeds.rb
@@ -4,20 +4,23 @@ puts "DESTROY --------------"
 # Hpi.destroy_all
 Encounter.destroy_all
 User.destroy_all
-Icd10.destroy_all
 
-puts "ICD10 --------------"
-csv_text = File.path(Rails.root.join("db", "Section111ValidICD10-Jan2020.csv"))
-count = 1
-CSV.foreach(csv_text) do |row|
-  Icd10.create(code: row[0], description: row[1])
-  count += 1
-  if (count % 1000) == 0
-    puts count
+importICD10 = false
+
+if importICD10
+  Icd10.destroy_all
+  puts "ICD10 --------------"
+  csv_text = File.path(Rails.root.join("db", "Section111ValidICD10-Jan2020.csv"))
+  count = 1
+  CSV.foreach(csv_text) do |row|
+    Icd10.create(code: row[0], description: row[1])
+    count += 1
+    if (count % 1000) == 0
+      puts count
+    end
   end
+  puts "ICD10 import done " + count.to_s + " records"
 end
-puts "ICD10 import done " + count.to_s + " records"
-
 puts "admin --------------"
 #admin
 User.create!({ first_name: "Clark", last_name: "Johnson", email: "clarkandkathy@gmail.com", role: 4, password: "Admin1234" })
